requirements.txt
sounddevice
numpy
TTS==0.22.0
transformers==4.49.0  # 修復GPT2InferenceModel兼容性問題
openai-whisper
google-generativeai
torch==2.1.1  # XTTS 與 Whisper 都依賴它
pyaudio
pygame
cutlet
fugashi  # 日語文本處理庫
unidic-lite  # 日語詞典，支援 XTTS 日語語音合成
faster-whisper 
whisper-timestamped 
wavmark
g2pkk
bleak

